HIIT-030 rename: common/exception에 있는 클래스 exception로 이동
